refactor resource container
